feat: create folder component in routess and move all componentsm
