+++ src/app/routes/components/route/route.component.scss
