+++ src/app/routes/components/routes-page/routes-page.component.html
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 <button tuiButton (click)="create()">Create</button>
 
 <tui-expand [expanded]="isCreating()">
@@ -18,3 +19,7 @@
     }
   }
 </div>
+=======
+<tra-routes-form />
+<tra-routes-list />
+>>>>>>> 4214b3d (feat: create folder component in routess and move all componentsm)
