--- src/app/routes/routes-page/routes-page.component.html
@@ -1,2 +1,0 @@
-<tra-routes-form />
-<tra-routes-list />
