+++ src/app/routes/shared/route-information/route-information.component.html
@@ -1,0 +1,1 @@
+<p>route-information works!</p>
