+++ src/app/routes/shared/route-information/route-information.component.ts
@@ -1,0 +1,13 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'tra-route-information',
+  standalone: true,
+  imports: [],
+  templateUrl: './route-information.component.html',
+  styleUrl: './route-information.component.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class RouteInformationComponent {
+
+}
